+++ go/apps/tao_launch/tao_launch.go
@@ -119,7 +119,6 @@ import (
 	"os/signal"
 	"path"
 	"regexp"
-	"strconv"
 	"strings"
 	"syscall"
 	"text/tabwriter"
@@ -130,7 +129,6 @@ import (
 	"github.com/jlmucb/cloudproxy/go/tao/auth"
 	"github.com/jlmucb/cloudproxy/go/util"
 	"github.com/jlmucb/cloudproxy/go/util/options"
-	"golang.org/x/crypto/ssh/terminal"
 )
 
 var opts = []options.Option{
@@ -565,58 +563,6 @@ func defaultSIGTTIN() {
 func dropSIGTTIN() {
 	fmt.Fprintf(noise, "[dropping SIGTTIN]\n")
 	signal.Notify(discard, syscall.SIGTTIN)
-}
-
-func isCtty(fd int) bool {
-	// One would hope there was a simple way to figure out what our controlling
-	// tty is. Or at least check if stdin is coming from our controlling tty (as
-	// opposed to just any old terminal). Alas, the infuriating morass that
-	// passes for job control provides no such ability. Of no help:
-	// stat(/dev/tty), readlink(/dev/fd/0), open(/dev/stdin), open(/dev/tty),
-	// cat(/proc/self/fdinfo/0), stat(/proc/self/fd/0), ioctl, tty_ioctl, TIOC*,
-	// anything to do with sid, pgrp, pgid, /dev/console, the tty command, $TTY,
-	// $SSH_TTY. Since I am on a mission, I delved into the source for /bin/ps
-	// to discover /proc/self/stat contains the major/minor numbers for the
-	// controlling tty. And stat(stdin).Rdev provides the same info. If they
-	// differ, I'm going to conclude -- oh so tentatively -- that stdin is NOT
-	// our controlling tty. If they match, or anything goes wrong, we will
-	// assume that stdin, if it is a terminal, is our ctty.
-	if !terminal.IsTerminal(fd) {
-		return false
-	}
-	var s syscall.Stat_t
-	err := syscall.Fstat(fd, &s)
-	if err != nil {
-		fmt.Fprintf(noise, "[warning: fstat(%d) failed: %v]\n", fd, err)
-		return true
-	}
-	name := "/proc/self/stat"
-	f, err := os.Open(name)
-	if err != nil {
-		fmt.Fprintf(noise, "[warning: open(%q) failed: %v]\n", name, err)
-		return true
-	}
-	b, err := ioutil.ReadAll(f)
-	if err != nil {
-		fmt.Fprintf(noise, "[warning: read(%q) failed: %v]\n", name, err)
-		return true
-	}
-	a := strings.Split(string(b), " ")
-	tty_nr := 6
-	if len(a) <= tty_nr {
-		fmt.Fprintf(noise, "[warning: read(%q) borked: only %d fields]\n", name, len(a))
-		return true
-	}
-	ctty, err := strconv.Atoi(a[tty_nr])
-	if err != nil {
-		fmt.Fprintf(noise, "[warning: read(%q) borked: tty_nr = %v]\n", name, a[tty_nr])
-		return true
-	}
-	if uint64(ctty) != s.Rdev {
-		fmt.Fprintf(noise, "[warning: stdin is a tty, but not ctty]\n")
-		return false
-	}
-	return true
 }
 
 func isForeground() bool {
